requirements.txt
discord.py==1.3.4
praw==5.2.0
psycopg2_binary==2.7.4
psycopg2==2.7.4
